pyproject.toml
[tool.poetry]
name = "stack-sparrow"
version = "0.1.0"
description = ""
authors = ["Akhil Ravidas <ar@mod0.ai>"]
readme = "README.md"
packages = [
    {include = "sparrow"},
]

[tool.poetry.dependencies]
python = "^3.11"
gitpython = "^3.1.40"
unidiff = "^0.7.5"
openai = "^1.1.1"
tiktoken = "^0.5.1"
sentry-sdk = "^1.34.0"
jinja2 = "^3.1.2"
typer = {extras = ["all"], version = "^0.9.0"}
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sparrow = "sparrow.main:app"